+++ actions/errActions.ts
@@ -3,6 +3,7 @@ import sendEmail from "@/app/api/utils/sendMail";
 
 export default async function sendErr(e: Error) {
   "use server";
+  if(process.env.NODE_ENV !== "production") return;
   await sendEmail({
     subject: "Error Caused",
     text: `Error caused with a client: